[#13] Refactor: 예보 파일 / 프로그레스바 파일 분리
- 예보 파일은 렌더링만, 프로그레스바는 프로그레스바 기능만 확실히 하기 위해 파일을 분리했다.
